Adventurer Game Object and CameraController Script
For easy debugging purposes, an Adventurer game object was added as well as additional scripts. The Adventurer utilizes the NavMesh by mouse clicking the camera space and sets the AI agent's destination.
A Gizmo was added to see where the AI agent's destination is.

CameraController Script was added to follow the Adventurer.