Early Implementation of Chests
Chest Prefab has been edited and now has a 2D box collider. Chests can be detected by the Objective Detector Script which is attached to the child of the Adventurer.

Once a chest has been detected by the Objective Detector'sTriggerEnter2D. The position of the chest is stored onto a list in the Adventurer's Controller for future uses.

Chest animations have been created.